New funny project for exercising purposes
Made a script that manages To-do-lists from command line. For now it just lets you create empty lists and name them. These lists are stored as txt files but I want to implement a way to store all of the lists in a single txt file or at maximum 2 txt files, one for storing position and the other for storage of data.